// frontend/src/shared/api-client/models/McpBindingCreateIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type McpBindingCreateIn = {
    allowed_tools?: Array<string>;
    config?: Record<string, any>;
    reference: string;
    source: 'builtin' | 'url' | 'package';
};